ISSUE #5310 - rename stuff
