+++ frontend/src/v5/ui/controls/resizableTableContext/resizableTableCell/resizableTableCell.styles.ts
